+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 2.2.0-dev
 - raised minimal version of PHP to 7.3
 - deprecated some getters and setters of BookControl and BookPage
+- made BookPage::$slug and BookPage::$title writable
 
 Version 2.1.1
 - re-added support for Latte 2.5